Remove diagnostics channel based log and bring back the old one
